recipes-iec61850: add libiec61850 and open-server
This commit adds the default conf/layer.conf for an optional seapath
layer, as well as 2 recipes within this layer. The 2 recipes contain the
libiec61850 library from MZ- Automation, as well as an IEC61850 open
server implementation. Both can be used for testing and platform
validation of the IEC 61850 protocol and real time requirements.

Signed-off-by: Robin Massink <robin.massink@alliander.com>